requirements.txt
aiohttp
pika
beautifulsoup4
python-dotenv
